refactor: move see all into packages
