+++ app/proguard-rules.pro
@@ -82,8 +82,8 @@
     public static *** i(...);
 }
 
--keep enum com.madrid.presentation.viewModel.seeAll.movies.SeeAllMoviesType { *; }
--keep enum com.madrid.presentation.viewModel.seeAll.tvShows.SeeAllTvShowType { *; }
+-keep enum com.madrid.presentation.viewModel.seeAll.movies.factory.SeeAllMoviesType { *; }
+-keep enum com.madrid.presentation.viewModel.seeAll.tvShows.factory.SeeAllTvShowType { *; }
 -keep enum com.madrid.presentation.viewModel.libraryViewModel.viewAll.factory.ViewAllType { *; }
