+++ presentation/src/main/java/com/madrid/presentation/screens/homeScreen/layout/MoviesLayout.kt
@@ -33,7 +33,7 @@ import com.madrid.presentation.component.MovioPager
 import com.madrid.presentation.component.movioCards.MovioVerticalCard
 import com.madrid.presentation.navigation.Destinations
 import com.madrid.presentation.navigation.LocalNavController
-import com.madrid.presentation.viewModel.seeAll.movies.SeeAllMoviesType
+import com.madrid.presentation.viewModel.seeAll.movies.factory.SeeAllMoviesType
 import com.madrid.presentation.viewModel.shared.MediaUiState
 import kotlinx.coroutines.flow.distinctUntilChanged
