+++ presentation/src/main/java/com/madrid/presentation/viewModel/seeAll/movies/strategy/SeeAllMoreRecommendedMovies.kt
@@ -1,24 +1,24 @@
-package com.madrid.presentation.viewModel.seeAll.movies
+package com.madrid.presentation.viewModel.seeAll.movies.strategy
 
 import com.madrid.domain.entity.Movie
 import com.madrid.domain.usecase.movie.FilterMoviesByCategoryUseCase
-import com.madrid.domain.usecase.movie.GetTopRatedMoviesUseCase
+import com.madrid.domain.usecase.search.GetRecommendedMovieUseCase
 
-class SeeAllTopRatedMovies(
-    val getTopRatedMoviesUseCase: GetTopRatedMoviesUseCase,
+class SeeAllMoreRecommendedMovies(
+    val getRecommendedMovieUseCase: GetRecommendedMovieUseCase,
     val filterMoviesByCategoryUseCase: FilterMoviesByCategoryUseCase
 ) : SeeAllMoviesStrategy {
     override fun getTitle(): String {
-        return "Top Rating"
+        return "More Recommended"
     }
 
     override suspend fun getAllMovies(page: Int): List<Movie> {
-        return getTopRatedMoviesUseCase(page)
+        return getRecommendedMovieUseCase(page)
     }
 
     override suspend fun getMoviesBasedOnCategory(categoryId: Int, page: Int): List<Movie> {
         val movies = getAllMovies(page)
-        return filterMoviesByCategoryUseCase(movies, categoryId)
+        return filterMoviesByCategoryUseCase(movies,categoryId)
     }
 
 }
