+++ presentation/src/main/java/com/madrid/presentation/viewModel/seeAll/movies/strategy/SeeAllNowPlayingMovies.kt
@@ -1,4 +1,4 @@
-package com.madrid.presentation.viewModel.seeAll.movies
+package com.madrid.presentation.viewModel.seeAll.movies.strategy
 
 import com.madrid.domain.entity.Movie
 import com.madrid.domain.usecase.movie.FilterMoviesByCategoryUseCase
