+++ presentation/src/main/java/com/madrid/presentation/viewModel/seeAll/movies/strategy/SeeAllTopRatedMovies.kt
@@ -1,19 +1,19 @@
-package com.madrid.presentation.viewModel.seeAll.movies
+package com.madrid.presentation.viewModel.seeAll.movies.strategy
 
 import com.madrid.domain.entity.Movie
 import com.madrid.domain.usecase.movie.FilterMoviesByCategoryUseCase
-import com.madrid.domain.usecase.movie.GetUpcomingMovieUseCase
+import com.madrid.domain.usecase.movie.GetTopRatedMoviesUseCase
 
-class SeeAllUpComingMovies(
-    val getUpcomingMovieUseCase: GetUpcomingMovieUseCase,
+class SeeAllTopRatedMovies(
+    val getTopRatedMoviesUseCase: GetTopRatedMoviesUseCase,
     val filterMoviesByCategoryUseCase: FilterMoviesByCategoryUseCase
 ) : SeeAllMoviesStrategy {
     override fun getTitle(): String {
-        return "Upcoming"
+        return "Top Rating"
     }
 
     override suspend fun getAllMovies(page: Int): List<Movie> {
-        return getUpcomingMovieUseCase(page)
+        return getTopRatedMoviesUseCase(page)
     }
 
     override suspend fun getMoviesBasedOnCategory(categoryId: Int, page: Int): List<Movie> {
