+++ presentation/src/main/java/com/madrid/presentation/viewModel/seeAll/tvShows/SeeAllTVShowsViewModel.kt
@@ -12,6 +12,7 @@ import com.madrid.presentation.pagination.SeeAllSeriesPagingSource
 import com.madrid.presentation.pagination.SeeAllSeriesWithGenrePagingSource
 import com.madrid.presentation.viewModel.shared.parser.formatRate
 import com.madrid.presentation.viewModel.base.BaseViewModel
+import com.madrid.presentation.viewModel.seeAll.tvShows.strategy.SeeAllTVShowsStrategy
 import com.madrid.presentation.viewModel.shared.CategoryUiState
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
