--- presentation/src/main/java/com/madrid/presentation/viewModel/seeAll/tvShows/Titles.kt
@@ -1,5 +1,0 @@
-package com.madrid.presentation.viewModel.seeAll.tvShows
-
-enum class Titles {
-
-}
