+++ presentation/src/main/java/com/madrid/presentation/viewModel/seeAll/tvShows/factory/SeeAllTVShowsFactory.kt
@@ -1,10 +1,15 @@
-package com.madrid.presentation.viewModel.seeAll.tvShows
+package com.madrid.presentation.viewModel.seeAll.tvShows.factory
 
 import com.madrid.domain.usecase.series.FilterSeriesByCategoryUseCase
 import com.madrid.domain.usecase.series.GetAiringTodaySeriesUseCase
 import com.madrid.domain.usecase.series.GetOnAirSeriesUseCase
 import com.madrid.domain.usecase.series.GetRecommendedSeriesUseCase
 import com.madrid.domain.usecase.series.GetTopRatedSeriesUseCase
+import com.madrid.presentation.viewModel.seeAll.tvShows.strategy.SeeAllAiringTodayTvShow
+import com.madrid.presentation.viewModel.seeAll.tvShows.strategy.SeeAllOnTvShow
+import com.madrid.presentation.viewModel.seeAll.tvShows.strategy.SeeAllRecommendedTVShow
+import com.madrid.presentation.viewModel.seeAll.tvShows.strategy.SeeAllTVShowsStrategy
+import com.madrid.presentation.viewModel.seeAll.tvShows.strategy.SeeAllTopRatingTVShows
 import javax.inject.Inject
 
 class SeeAllTVShowsFactory @Inject constructor(
