+++ presentation/src/main/java/com/madrid/presentation/viewModel/seeAll/tvShows/strategy/SeeAllRecommendedTVShow.kt
@@ -1,22 +1,22 @@
-package com.madrid.presentation.viewModel.seeAll.tvShows
+package com.madrid.presentation.viewModel.seeAll.tvShows.strategy
 
 import com.madrid.domain.entity.Series
 import com.madrid.domain.usecase.series.FilterSeriesByCategoryUseCase
-import com.madrid.domain.usecase.series.GetOnAirSeriesUseCase
+import com.madrid.domain.usecase.series.GetRecommendedSeriesUseCase
 
-class SeeAllOnTvShow(
-    private val getOnAirSeriesUseCase: GetOnAirSeriesUseCase,
+class SeeAllRecommendedTVShow(
+    private val getRecommendedSeriesUseCase: GetRecommendedSeriesUseCase,
     private val filterSeriesByCategoryUseCase: FilterSeriesByCategoryUseCase
 ) : SeeAllTVShowsStrategy {
     override fun getTitle(): String {
-        return "On TV"
+        return "More Recommended"
     }
 
     override suspend fun getAllTvShows(page: Int): List<Series> {
-        return getOnAirSeriesUseCase(page)
+        return getRecommendedSeriesUseCase(page)
     }
 
-    override suspend fun getTvShowsBasedOnCategory(categoryId: Int, page: Int): List<Series> {
+    override suspend fun getTvShowsBasedOnCategory(categoryId: Int,page: Int): List<Series> {
         val tvShows = getAllTvShows(page)
         return filterSeriesByCategoryUseCase(series = tvShows, category = categoryId)
     }
